refactor: create competition-groups module
