+++ src/components/forms/player/players-filter-form.tsx
@@ -12,7 +12,7 @@ import { PlayersPositionCombo } from '@/components/selects/player-positions-comb
 import { TeamsCombo } from '@/components/selects/teams-combo'
 import { PlayersFiltersDto } from '@/modules/players/types'
 import { TeamBasicDataDto } from '@/modules/teams/types'
-import { CompetitionGroupBasicDataDto } from '@/types/competition-groups'
+import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
 import { CompetitionBasicDataDto } from '@/types/competitions'
 import { CountryDto } from '@/types/countries'
 import { PlayerPositionDto } from '@/types/player-positions'
