+++ src/components/forms/team/teams-filter-form.tsx
@@ -10,7 +10,7 @@ import { CompetitionsCombo } from '@/components/selects/competitions-combo'
 import { CountriesCombo } from '@/components/selects/countries-combo'
 import { RegionsCombo } from '@/components/selects/regions-combo'
 import { ClubBasicDataDto, ClubsFiltersDto } from '@/modules/clubs/types'
-import { CompetitionGroupBasicDataDto } from '@/types/competition-groups'
+import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
 import { CompetitionBasicDataDto } from '@/types/competitions'
 import { CountryDto } from '@/types/countries'
 import { RegionDto } from '@/types/regions'
