+++ src/pages/teams/create.tsx
@@ -4,10 +4,10 @@ import { serverSideTranslations } from 'next-i18next/serverSideTranslations'
 import { CreateTeamForm } from '@/components/forms/team/create-team'
 import { Loader } from '@/components/loader/loader'
 import { PageHeading } from '@/components/page-heading/page-heading'
-import { useCompetitionGroupsList } from '@/lib/competition-groups'
 import { useCompetitionsList } from '@/lib/competitions'
 import { withSessionSsr } from '@/lib/session'
 import { useClubsList } from '@/modules/clubs/hooks'
+import { useCompetitionGroupsList } from '@/modules/competition-groups/hooks'
 import { useCreateTeam } from '@/modules/teams/hooks'
 import { redirectToLogin } from '@/utils/redirect-to-login'
