+++ src/services/api/methods/competition-groups.ts
@@ -1,0 +1,9 @@
+import { CompetitionGroupBasicDataDto } from '@/modules/competition-groups/types'
+import { TModuleName } from '@/services/api/modules'
+
+import { getDataList } from './helpers'
+
+const moduleName: TModuleName = 'competition-groups'
+
+export const getCompetitionGroupsList = () =>
+  getDataList<CompetitionGroupBasicDataDto>(moduleName)
